Clean up private repo resolution
Motivation:

The current way to resolve private publish repo is a mess:
* decide based on `pushToPrivateNexus` sbt settings to publish on
public or private
* populate `publishTo` settings based on envar
* populate `credentials` settings based on properties file

Configuring the env var every time we want to publish is a pita.
The host value is duplicated in the envar and the props file. We could
just use the props file, the only missing information is the scheme.

Modification:

* Drop envar
* Add scheme in props file
* Resolve `publishTo` settings based on props file

Result:

Easier publish process (no more envar)