Enforce Java 8 API
Motivation:

We want to be able to compile with more modern Java version and still be compatible with Java 8, as long as source code is compatible.

In particular, we don't want to crash with NoSuchMethodError because we would be linking to the new ByteBuffer/Buffer overloads introduced in Java 8.

Modifications:

* enable `-release` scalac option when building this plugin
* enable `-release` scalac option for projects importing this plugin

Result:

This plugin can now compile projects with Java 11 and still emit bytecode compatible with Java 8.